modified code
