+++ src/exception/WrongWordLengthException.java
@@ -1,2 +1,4 @@
+package exception;
+
 public class WrongWordLengthException extends Exception {
 }
